fix: game crashed because the wrong event
