+++ common/src/main/kotlin/com/rpmtw/rpmtw_platform_mod/config/ConfigObject.kt
@@ -61,8 +61,8 @@ class ConfigObject : ConfigData {
         @ConfigEntry.Gui.RequiresRestart
         var loadTranslateResourcePack = true
             get() {
-                return if (GameLanguage.getMinecraft() == GameLanguage.English) {
-                    false
+                return if (GameLanguage.getMinecraft() == GameLanguage.TraditionalChinese) {
+                    true
                 } else {
                     field
                 }
