+++ common/src/main/kotlin/com/rpmtw/rpmtw_platform_mod/handlers/EventHandler.kt
@@ -8,7 +8,6 @@ import dev.architectury.event.events.client.ClientTooltipEvent
 
 object EventHandler {
     fun handle() {
-        ClientLifecycleEvent.CLIENT_SETUP.register(OnClientSetup())
         ClientLifecycleEvent.CLIENT_STOPPING.register(OnClientStopping())
         ClientLifecycleEvent.CLIENT_STARTED.register(OnClientStarted())
         ClientGuiEvent.INIT_POST.register(OnGuiInitPost())
